fix: Removed key sengrid from .env file and fixed bugs
