+++ .env
@@ -1,4 +1,4 @@
 DATABASE_URL="postgresql://dev_admin:dev_password@localhost:5433/postgres?schema=public"
 JWT_SECRET=fdf8s8fs8fyfg8ffsfgsf8fsf8vwvywfefknflkdsf8f
 COOKIE_NAME=checklisty_cookie
-SENDGRID_API_KEY=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
+SENDGRID_API_KEY=i can`t add key here bacause SendGrid will block my account
