+++ .gitignore
@@ -27,7 +27,6 @@ yarn-error.log*
 # local env files
 .env
 .env*.local
-.env
 
 # vercel
 .vercel
