+++ app/(auth)/sign-in/page.tsx
@@ -1,4 +1,5 @@
 import { SignIn } from '@/components'
+// import SignIn from '../../../components/SignIn'
 
 export default function Home() {
     return (
