+++ app/layout.tsx
@@ -1,22 +1,33 @@
+// 'use client'
 import '@/styles/globals.css'
 
-// import { NavBar } from '@/components'
+import { cookies } from 'next/headers'
+
+import { NavBar } from '@/components'
+import { delay } from '@/lib/async'
+import { getUserFromCookie } from '@/lib/auth'
 
 export const metadata = {
     title: 'Checklisty',
     description: 'Checklisty is a checklist app for your daily tasks.',
 }
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+const getData = async () => {
+    await delay(2300)
+    const user = await getUserFromCookie(cookies())
+    return user
+}
+
+export default async function RootLayout({ children }: { children: React.ReactNode }) {
+    const user = await getData()
+    console.log('render layout')
     return (
         <html lang="en">
             <body>
-                {/* <main className="relative z-0 bg-primary"> */}
-                    {/* <header> */}
-                        {/* <NavBar /> */}
-                    {/* </header> */}
-                    {children}
-                {/* </main> */}
+                <header>
+                    <NavBar user={user?.name} />
+                </header>
+                <main className="relative z-0 bg-primary">{children}</main>
             </body>
         </html>
     )
